refactor : repackaging tablegroup
테이블그룹 관련 코드 리패키징
