+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -1,11 +1,11 @@
-package kitchenpos.application;
+package kitchenpos.tablegroup.application;
 
 import kitchenpos.dao.OrderDao;
 import kitchenpos.table.dao.OrderTableDao;
-import kitchenpos.dao.TableGroupDao;
+import kitchenpos.tablegroup.dao.TableGroupDao;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
